requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
Django==4.2.5
django-paypal==2.0
idna==3.4
paypalrestsdk==1.13.1
pycparser==2.21
pyOpenSSL==23.2.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.5
